[CCUBE-2118][RL] Refactor thumbnail styles
